First temporal commit
This is a temporal Github repository in order to converge to a Public
Fork of CS_CityScope_GAMA once the war would be finished.

This repo needs to be private and the forking of a public repo doesn’t
allow to make the fork private. That’s why we are using this temporal
solution